Modularização da conexão Wifi
