+++ APIconection/http_server.c
@@ -1,0 +1,78 @@
+#include "pico/cyw43_arch.h"
+#include "pico/stdlib.h"
+#include "lwip/tcp.h"
+#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <ctype.h>
+//Módulos adicionados
+#include "CREDENTIALS.h"
+
+
+
+#define LED_PIN 12          // Define o pino do LED
+//#define WIFI_SSID ""  // Substitua pelo nome da sua rede Wi-Fi
+//#define WIFI_PASS "" // Substitua pela senha da sua rede Wi-Fi
+
+// Buffer para respostas HTTP
+#define HTTP_RESPONSE "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\n\r\n" \
+                      "<!DOCTYPE html><html><body>" \
+                      "<h1>Controle do LED</h1>" \
+                      "<p><a href=\"/led/on\">Ligar LED</a></p>" \
+                      "<p><a href=\"/led/off\">Desligar LED</a></p>" \
+                      "</body></html>\r\n"
+
+
+
+// Função de callback para processar requisições HTTP
+static err_t http_callback(void *arg, struct tcp_pcb *tpcb, struct pbuf *p, err_t err) {
+    if (p == NULL) {
+        // Cliente fechou a conexão
+        tcp_close(tpcb);
+        return ERR_OK;
+    }
+
+    // Processa a requisição HTTP
+    char *request = (char *)p->payload;
+
+    if (strstr(request, "GET /led/on")) {
+        gpio_put(LED_PIN, 1);  // Liga o LED
+    } else if (strstr(request, "GET /led/off")) {
+        gpio_put(LED_PIN, 0);  // Desliga o LED
+    }
+
+    // Envia a resposta HTTP
+    tcp_write(tpcb, HTTP_RESPONSE, strlen(HTTP_RESPONSE), TCP_WRITE_FLAG_COPY);
+
+    // Libera o buffer recebido
+    pbuf_free(p);
+
+    return ERR_OK;
+}
+
+// Callback de conexão: associa o http_callback à conexão
+static err_t connection_callback(void *arg, struct tcp_pcb *newpcb, err_t err) {
+    tcp_recv(newpcb, http_callback);  // Associa o callback HTTP
+    return ERR_OK;
+}
+
+// Função de setup do servidor TCP
+static void start_http_server(void) {
+    struct tcp_pcb *pcb = tcp_new();
+    if (!pcb) {
+        printf("Erro ao criar PCB\n");
+        return;
+    }
+
+    // Liga o servidor na porta 80
+    if (tcp_bind(pcb, IP_ADDR_ANY, 80) != ERR_OK) {
+        printf("Erro ao ligar o servidor na porta 80\n");
+        return;
+    }
+
+    pcb = tcp_listen(pcb);  // Coloca o PCB em modo de escuta
+    tcp_accept(pcb, connection_callback);  // Associa o callback de conexão
+
+    printf("Servidor HTTP rodando na porta 80...\n");
+}
+
